move all integer_adm code to single file
