+++ libvmaf/src/meson.build
@@ -158,7 +158,6 @@ libvmaf_rc_feature_sources = [
   feature_src_dir + 'alias.c',
   feature_src_dir + 'float_adm.c',
   feature_src_dir + 'integer_adm.c',
-  feature_src_dir + 'integer_adm_function.c',
   feature_src_dir + 'feature_collector.c',
   feature_src_dir + 'float_psnr.c',
   feature_src_dir + 'float_motion.c',
